feat: implement widthdrawal request, add to google calendar, create dummy data and fix: some problems

+++ src/main/java/com/auhpp/event_management/entity/Coupon.java
@@ -46,7 +46,7 @@ public class Coupon {
     @Column(nullable = false)
     private Double value;
 
-    private Integer maxDiscountAmount;
+    private Double maxDiscountAmount;
 
     private Integer maximumUsage;
